refactoring stats leaderboard 2
