+++ lib/screens/Stats Leaderboard/leaderboard.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/material.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:my_softball_team/globals.dart' as globals;
+
+class Leaderboard extends StatefulWidget {
+  @override
+  _LeaderboardState createState() => _LeaderboardState();
+}
+
+class _LeaderboardState extends State<Leaderboard> {
+  CollectionReference players = Firestore.instance
+    .collection("Teams")
+    .document(globals.teamName)
+    .collection("Players");
+
+  @override
+  Widget build(BuildContext context) {
+    return StreamBuilder<QuerySnapshot>(
+      stream: players.snapshots(),
+      builder: (context, snapshot) {
+        if (!snapshot.hasData) {
+          return Center(
+            child: CircularProgressIndicator(),
+          );
+        } else {
+          final player = snapshot;
+          return Expanded(
+            child: ListView.builder(
+              itemCount: player.data.documents.length,
+              itemBuilder: (context, index) {
+                return Padding(
+                  padding: const EdgeInsets.all(16.0),
+                  child: Material(
+                    elevation: 2.0,
+                    borderRadius: BorderRadius.all(Radius.circular(10.0)),
+                    child: Padding(
+                      padding: const EdgeInsets.all(8.0),
+                      child: Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        children: <Widget>[
+                          Text((index + 1).toString()),
+                          Text(player.data.documents[index]['PlayerName']),
+                          //Text(statSelection != null ? player.data.documents[index][statSelection] : ""),
+                        ],
+                      ),
+                    ),
+                  ),
+                );
+              },
+            ),
+          );
+        }
+      },
+    );
+  }
+}
